Extract auth activity backup and API key domains
