+++ crates/clickhousectl/src/cloud/client.rs
@@ -577,34 +577,6 @@ impl CloudClient {
         Self::unwrap_response(response)
     }
 
-    // Backup endpoints (delegated to library client)
-    pub async fn list_backups(
-        &self,
-        org_id: &str,
-        service_id: &str,
-    ) -> Result<Vec<clickhouse_cloud_api::models::Backup>> {
-        let response = self
-            .api()
-            .backup_get_list(org_id, service_id)
-            .await
-            .map_err(|e| self.convert_error_for_organization(e, org_id))?;
-        Self::unwrap_response(response)
-    }
-
-    pub async fn get_backup(
-        &self,
-        org_id: &str,
-        service_id: &str,
-        backup_id: &str,
-    ) -> Result<clickhouse_cloud_api::models::Backup> {
-        let response = self
-            .api()
-            .backup_get(org_id, service_id, backup_id)
-            .await
-            .map_err(|e| self.convert_error_for_organization(e, org_id))?;
-        Self::unwrap_response(response)
-    }
-
     // Update service
     pub async fn update_service(
         &self,
@@ -883,142 +855,6 @@ impl CloudClient {
             status: response.status,
             request_id: response.request_id,
         })
-    }
-
-    // Phase 5 - API Key endpoints (delegated to library client)
-    pub async fn list_api_keys(
-        &self,
-        org_id: &str,
-    ) -> Result<Vec<clickhouse_cloud_api::models::ApiKey>> {
-        let response = self
-            .api()
-            .openapi_key_get_list(org_id)
-            .await
-            .map_err(|e| self.convert_error_for_organization(e, org_id))?;
-        Self::unwrap_response(response)
-    }
-
-    pub async fn create_api_key(
-        &self,
-        org_id: &str,
-        request: &clickhouse_cloud_api::models::ApiKeyPostRequest,
-    ) -> Result<clickhouse_cloud_api::models::ApiKeyPostResponse> {
-        let response = self
-            .api()
-            .openapi_key_create(org_id, request)
-            .await
-            .map_err(|e| self.convert_error_for_organization(e, org_id))?;
-        Self::unwrap_response(response)
-    }
-
-    pub async fn get_api_key(
-        &self,
-        org_id: &str,
-        key_id: &str,
-    ) -> Result<clickhouse_cloud_api::models::ApiKey> {
-        let response = self
-            .api()
-            .openapi_key_get(org_id, key_id)
-            .await
-            .map_err(|e| self.convert_error_for_organization(e, org_id))?;
-        Self::unwrap_response(response)
-    }
-
-    pub async fn update_api_key(
-        &self,
-        org_id: &str,
-        key_id: &str,
-        request: &clickhouse_cloud_api::models::ApiKeyPatchRequest,
-    ) -> Result<clickhouse_cloud_api::models::ApiKey> {
-        let response = self
-            .api()
-            .openapi_key_update(org_id, key_id, request)
-            .await
-            .map_err(|e| self.convert_error_for_organization(e, org_id))?;
-        Self::unwrap_response(response)
-    }
-
-    pub async fn delete_api_key(&self, org_id: &str, key_id: &str) -> Result<DeleteResponse> {
-        let response = self
-            .api()
-            .openapi_key_delete(org_id, key_id)
-            .await
-            .map_err(|e| self.convert_error_for_organization(e, org_id))?;
-        Ok(DeleteResponse {
-            status: response.status,
-            request_id: response.request_id,
-        })
-    }
-
-    pub async fn delete_api_key_if_exists(
-        &self,
-        org_id: &str,
-        key_id: &str,
-    ) -> Result<Option<DeleteResponse>> {
-        match self.api().openapi_key_delete(org_id, key_id).await {
-            Ok(response) => Ok(Some(DeleteResponse {
-                status: response.status,
-                request_id: response.request_id,
-            })),
-            Err(clickhouse_cloud_api::Error::Api { status: 404, .. }) => Ok(None),
-            Err(error) => Err(self.convert_error_for_organization(error, org_id)),
-        }
-    }
-
-    // Phase 6 - Activity endpoints
-    pub async fn list_activities(
-        &self,
-        org_id: &str,
-        from_date: Option<&str>,
-        to_date: Option<&str>,
-    ) -> Result<Vec<clickhouse_cloud_api::models::Activity>> {
-        let response = self
-            .api()
-            .activity_get_list(org_id, from_date, to_date)
-            .await
-            .map_err(|e| self.convert_error_for_organization(e, org_id))?;
-        Self::unwrap_response(response)
-    }
-
-    pub async fn get_activity(
-        &self,
-        org_id: &str,
-        activity_id: &str,
-    ) -> Result<clickhouse_cloud_api::models::Activity> {
-        let response = self
-            .api()
-            .activity_get(org_id, activity_id)
-            .await
-            .map_err(|e| self.convert_error_for_organization(e, org_id))?;
-        Self::unwrap_response(response)
-    }
-
-    // Backup Config endpoints (delegated to library client)
-    pub async fn get_backup_config(
-        &self,
-        org_id: &str,
-        service_id: &str,
-    ) -> Result<clickhouse_cloud_api::models::BackupConfiguration> {
-        let response = self
-            .api()
-            .backup_configuration_get(org_id, service_id)
-            .await
-            .map_err(|e| self.convert_error_for_organization(e, org_id))?;
-        Self::unwrap_response(response)
-    }
-
-    pub async fn update_backup_config(
-        &self,
-        org_id: &str,
-        service_id: &str,
-        request: &clickhouse_cloud_api::models::BackupConfigurationPatchRequest,
-    ) -> Result<clickhouse_cloud_api::models::BackupConfiguration> {
-        let response = self
-            .api()
-            .backup_configuration_update(org_id, service_id, request)
-            .await
-            .map_err(|e| self.convert_error_for_organization(e, org_id))?;
-        Self::unwrap_response(response)
     }
 
     // ClickPipe endpoints (delegated to library client)
