+++ crates/clickhousectl/src/cloud/commands.rs
@@ -1,12 +1,11 @@
+use crate::cloud::api_keys::{cleanup_service_query_key, service_query_key_cleanup};
 use crate::cloud::client::{CloudClient, CloudError};
 use crate::cloud::credentials;
 use crate::cloud::output::{ABSENT, or_absent, print_human};
 use crate::cloud::shared::{parse_serde_enum, parse_tags, resolve_org_id};
 use clickhouse_cloud_api::models::{
-    ApiKeyPatchRequest, ApiKeyPatchRequestState, ApiKeyPostRequest, ApiKeyPostRequestState,
-    AutoscalingMode, BackupConfigurationPatchRequest, InstancePrivateEndpointsPatch,
-    InstanceServiceQueryApiEndpointsPostRequest, InstanceTagsPatch, IpAccessListEntry,
-    IpAccessListPatch, OrganizationPatchPrivateEndpoint,
+    AutoscalingMode, InstancePrivateEndpointsPatch, InstanceServiceQueryApiEndpointsPostRequest,
+    InstanceTagsPatch, IpAccessListEntry, IpAccessListPatch, OrganizationPatchPrivateEndpoint,
     OrganizationPatchPrivateEndpointCloudprovider, OrganizationPatchPrivateEndpointRegion,
     OrganizationPatchRequest, OrganizationPrivateEndpointsPatch, ServicPrivateEndpointePostRequest,
     Service, ServiceEndpoint, ServiceEndpointChange, ServiceEndpointChangeProtocol,
@@ -15,7 +14,7 @@ use clickhouse_cloud_api::models::{
     ServicePostRequestProvider, ServicePostRequestRegion, ServicePostRequestReleasechannel,
     ServiceReplicaScalingPatchRequest, ServiceState, ServiceStatePatchRequestCommand,
 };
-use std::io::{IsTerminal, Write};
+use std::io::IsTerminal;
 use tabled::{Table, Tabled, settings::Style};
 
 /// Comma-joins the rendered items of a response list.
@@ -231,94 +230,6 @@ fn parse_org_private_endpoints_patch(
     }))
 }
 
-fn parse_api_key_hash_data(
-    key_id_hash: Option<&str>,
-    key_id_suffix: Option<&str>,
-    key_secret_hash: Option<&str>,
-) -> Result<Option<clickhouse_cloud_api::models::ApiKeyHashData>, Box<dyn std::error::Error>> {
-    match (key_id_hash, key_id_suffix, key_secret_hash) {
-        (None, None, None) => Ok(None),
-        (Some(key_id_hash), Some(key_id_suffix), Some(key_secret_hash)) => {
-            Ok(Some(clickhouse_cloud_api::models::ApiKeyHashData {
-                key_id_hash: key_id_hash.to_string(),
-                key_id_suffix: key_id_suffix.to_string(),
-                key_secret_hash: key_secret_hash.to_string(),
-            }))
-        }
-        _ => Err(
-            "pre-hashed API key input requires --hash-key-id, --hash-key-id-suffix, and --hash-key-secret together"
-                .into(),
-        ),
-    }
-}
-
-fn parse_ip_access_entries_lib(values: &[String]) -> Option<Vec<IpAccessListEntry>> {
-    (!values.is_empty()).then(|| {
-        values
-            .iter()
-            .map(|value| IpAccessListEntry {
-                source: value.clone(),
-                description: None,
-            })
-            .collect()
-    })
-}
-
-fn parse_uuid_list(
-    values: &[String],
-    field: &str,
-) -> Result<Vec<uuid::Uuid>, Box<dyn std::error::Error>> {
-    values
-        .iter()
-        .map(|s| {
-            uuid::Uuid::parse_str(s)
-                .map_err(|e| format!("invalid {} UUID '{}': {}", field, s, e).into())
-        })
-        .collect()
-}
-
-fn parse_api_key_state_post(
-    value: &str,
-) -> Result<ApiKeyPostRequestState, Box<dyn std::error::Error>> {
-    match value {
-        "enabled" => Ok(ApiKeyPostRequestState::Enabled),
-        "disabled" => Ok(ApiKeyPostRequestState::Disabled),
-        _ => Err(format!(
-            "invalid state: unknown value '{}', expected one of: enabled, disabled",
-            value
-        )
-        .into()),
-    }
-}
-
-fn parse_api_key_state_patch(
-    value: &str,
-) -> Result<ApiKeyPatchRequestState, Box<dyn std::error::Error>> {
-    match value {
-        "enabled" => Ok(ApiKeyPatchRequestState::Enabled),
-        "disabled" => Ok(ApiKeyPatchRequestState::Disabled),
-        _ => Err(format!(
-            "invalid state: unknown value '{}', expected one of: enabled, disabled",
-            value
-        )
-        .into()),
-    }
-}
-
-fn parse_expire_at(
-    value: &str,
-) -> Result<chrono::DateTime<chrono::Utc>, Box<dyn std::error::Error>> {
-    chrono::DateTime::parse_from_rfc3339(value)
-        .map(|dt| dt.with_timezone(&chrono::Utc))
-        .map_err(|e| {
-            format!(
-                "invalid expire_at '{}': expected ISO 8601 / RFC 3339 format (e.g. 2025-12-31T23:59:59Z): {}",
-                value, e
-            )
-            .into()
-        })
-}
-
 pub async fn org_list(client: &CloudClient, json: bool) -> Result<(), Box<dyn std::error::Error>> {
     let orgs = client.list_organizations().await?;
 
@@ -502,37 +413,6 @@ pub struct OrgUpdateOptions {
     pub name: Option<String>,
     pub remove_private_endpoints: Vec<String>,
     pub enable_core_dumps: Option<bool>,
-}
-
-#[derive(Default)]
-pub struct KeyCreateOptions {
-    pub name: String,
-    pub role_ids: Vec<String>,
-    pub expires_at: Option<String>,
-    pub state: Option<String>,
-    pub ip_allow: Vec<String>,
-    pub hash_key_id: Option<String>,
-    pub hash_key_id_suffix: Option<String>,
-    pub hash_key_secret: Option<String>,
-    pub org_id: Option<String>,
-}
-
-#[derive(Default)]
-pub struct KeyUpdateOptions {
-    pub name: Option<String>,
-    pub role_ids: Vec<String>,
-    pub expires_at: Option<String>,
-    pub state: Option<String>,
-    pub ip_allow: Vec<String>,
-    pub org_id: Option<String>,
-}
-
-#[derive(Default)]
-pub struct BackupConfigUpdateOptions {
-    pub backup_period_hours: Option<u32>,
-    pub backup_retention_period_hours: Option<u32>,
-    pub backup_start_time: Option<String>,
-    pub org_id: Option<String>,
 }
 
 /// Resolved horizontal-autoscaling fields for a service create/scale request.
@@ -733,68 +613,6 @@ fn build_org_update_request(
     })
 }
 
-fn build_api_key_create_request(
-    opts: &KeyCreateOptions,
-) -> Result<ApiKeyPostRequest, Box<dyn std::error::Error>> {
-    Ok(ApiKeyPostRequest {
-        name: opts.name.clone(),
-        expire_at: opts
-            .expires_at
-            .as_deref()
-            .map(parse_expire_at)
-            .transpose()?,
-        state: match opts.state.as_deref() {
-            Some(value) => parse_api_key_state_post(value)?,
-            None => ApiKeyPostRequestState::default(),
-        },
-        assigned_role_ids: parse_uuid_list(&opts.role_ids, "role_id")?,
-        ip_access_list: parse_ip_access_entries_lib(&opts.ip_allow).unwrap_or_default(),
-        hash_data: parse_api_key_hash_data(
-            opts.hash_key_id.as_deref(),
-            opts.hash_key_id_suffix.as_deref(),
-            opts.hash_key_secret.as_deref(),
-        )?,
-        #[cfg(feature = "deprecated-fields")]
-        roles: None,
-    })
-}
-
-fn build_api_key_update_request(
-    opts: &KeyUpdateOptions,
-) -> Result<ApiKeyPatchRequest, Box<dyn std::error::Error>> {
-    Ok(ApiKeyPatchRequest {
-        name: opts.name.clone(),
-        assigned_role_ids: if opts.role_ids.is_empty() {
-            None
-        } else {
-            Some(parse_uuid_list(&opts.role_ids, "role_id")?)
-        },
-        expire_at: opts
-            .expires_at
-            .as_deref()
-            .map(parse_expire_at)
-            .transpose()?,
-        state: opts
-            .state
-            .as_deref()
-            .map(parse_api_key_state_patch)
-            .transpose()?,
-        ip_access_list: parse_ip_access_entries_lib(&opts.ip_allow),
-        #[cfg(feature = "deprecated-fields")]
-        roles: None,
-    })
-}
-
-fn build_backup_config_update_request(
-    opts: &BackupConfigUpdateOptions,
-) -> BackupConfigurationPatchRequest {
-    BackupConfigurationPatchRequest {
-        backup_period_in_hours: opts.backup_period_hours.map(f64::from),
-        backup_retention_period_in_hours: opts.backup_retention_period_hours.map(f64::from),
-        backup_start_time: opts.backup_start_time.clone(),
-    }
-}
-
 /// The post-create hint showing how to query the new service.
 ///
 /// The hint is only useful with a real service id: an absent id would render a
@@ -930,65 +748,6 @@ fn service_delete_error(error: CloudError, force: bool, service_id: &str) -> Clo
     } else {
         error
     }
-}
-
-/// Return a query API key only when its exact resource and organization IDs
-/// were saved during provisioning. The boolean indicates that partial cleanup
-/// metadata must remain on disk because discarding it would lose the key ID.
-fn service_query_key_cleanup(
-    org_id: &str,
-    service_id: &str,
-) -> Result<(Option<String>, bool), Box<dyn std::error::Error>> {
-    let Some(key) = credentials::try_get_service_query_key(service_id)? else {
-        return Ok((None, false));
-    };
-    let Some(api_key_id) = key.api_key_id else {
-        eprintln!(
-            "Warning: the stored query key for service {service_id} predates exact management \
-             API key IDs; service deletion will continue without unsafe cloud key cleanup."
-        );
-        return Ok((None, false));
-    };
-    let Some(key_org_id) = key.organization_id else {
-        eprintln!(
-            "Warning: the stored query key for service {service_id} has a management API key ID \
-             but no provisioning organization; cloud key cleanup was skipped and the local \
-             record was retained."
-        );
-        return Ok((None, true));
-    };
-    if key_org_id != org_id {
-        return Err(format!(
-            "the stored query key for service {service_id} belongs to organization {key_org_id}, \
-             not {org_id}; refusing to delete either resource"
-        )
-        .into());
-    }
-    Ok((Some(api_key_id), false))
-}
-
-async fn cleanup_service_query_key(
-    client: &CloudClient,
-    org_id: &str,
-    service_id: &str,
-    api_key_id: Option<&str>,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let Some(api_key_id) = api_key_id else {
-        return Ok(());
-    };
-
-    client
-        .delete_api_key_if_exists(org_id, api_key_id)
-        .await
-        .map_err(|mut error| {
-            error.message = format!(
-                "failed to delete the auto-provisioned query API key for service \
-                 {service_id}: {}",
-                error.message
-            );
-            error
-        })?;
-    Ok(())
 }
 
 pub async fn service_delete(
@@ -2080,98 +1839,6 @@ pub async fn clickpipe_create_bigquery(
     Ok(())
 }
 
-pub async fn backup_list(
-    client: &CloudClient,
-    service_id: &str,
-    org_id: Option<&str>,
-    json: bool,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let org_id = resolve_org_id(client, org_id).await?;
-
-    let backups = client.list_backups(&org_id, service_id).await?;
-
-    if json {
-        println!("{}", serde_json::to_string_pretty(&backups)?);
-    } else {
-        if backups.is_empty() {
-            println!("No backups found");
-            return Ok(());
-        }
-        #[derive(Tabled)]
-        struct Row {
-            #[tabled(rename = "ID")]
-            id: String,
-            #[tabled(rename = "Status")]
-            status: String,
-            #[tabled(rename = "Size")]
-            size: String,
-            #[tabled(rename = "Created")]
-            created: String,
-        }
-        let rows: Vec<Row> = backups
-            .into_iter()
-            .map(|b| Row {
-                id: or_absent(b.id),
-                status: or_absent(b.status.as_ref()),
-                size: or_absent(b.size_in_bytes.map(format_bytes)),
-                created: or_absent(b.started_at.map(|at| at.to_rfc3339())),
-            })
-            .collect();
-        println!("{}", Table::new(rows).with(Style::markdown()));
-    }
-    Ok(())
-}
-
-pub async fn backup_get(
-    client: &CloudClient,
-    service_id: &str,
-    backup_id: &str,
-    org_id: Option<&str>,
-    json: bool,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let org_id = resolve_org_id(client, org_id).await?;
-
-    let backup = client.get_backup(&org_id, service_id, backup_id).await?;
-
-    if json {
-        println!("{}", serde_json::to_string_pretty(&backup)?);
-    } else {
-        print_human(&backup)?;
-    }
-    Ok(())
-}
-
-pub fn auth_interactive() -> Result<(), Box<dyn std::error::Error>> {
-    print!("API Key: ");
-    std::io::stdout().flush()?;
-    let mut api_key = String::new();
-    std::io::stdin().read_line(&mut api_key)?;
-    let api_key = api_key.trim().to_string();
-
-    if api_key.is_empty() {
-        return Err("API key cannot be empty".into());
-    }
-
-    print!("API Secret: ");
-    std::io::stdout().flush()?;
-    let api_secret = rpassword::read_password()?;
-
-    if api_secret.is_empty() {
-        return Err("API secret cannot be empty".into());
-    }
-
-    let mut creds = credentials::load_credentials().unwrap_or_default();
-    creds.api_key = Some(api_key);
-    creds.api_secret = Some(api_secret);
-    credentials::save_credentials(&creds)?;
-
-    println!(
-        "Credentials saved to {}",
-        credentials::credentials_path().display()
-    );
-    Ok(())
-}
-
 pub async fn service_update(
     client: &CloudClient,
     service_id: &str,
@@ -3086,334 +2753,6 @@ pub async fn invitation_delete(
     Ok(())
 }
 
-// =============================================================================
-// Phase 5 — API Key command handlers
-// =============================================================================
-
-pub async fn key_list(
-    client: &CloudClient,
-    org_id: Option<&str>,
-    json: bool,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let org_id = resolve_org_id(client, org_id).await?;
-
-    let keys = client.list_api_keys(&org_id).await?;
-
-    if json {
-        println!("{}", serde_json::to_string_pretty(&keys)?);
-    } else {
-        if keys.is_empty() {
-            println!("No API keys found");
-            return Ok(());
-        }
-        #[derive(Tabled)]
-        struct Row {
-            #[tabled(rename = "Name")]
-            name: String,
-            #[tabled(rename = "ID")]
-            id: String,
-            #[tabled(rename = "State")]
-            state: String,
-            #[tabled(rename = "Expires")]
-            expires: String,
-        }
-        let rows: Vec<Row> = keys
-            .into_iter()
-            .map(|k| Row {
-                name: or_absent(k.name.as_deref()),
-                id: or_absent(k.id),
-                state: or_absent(k.state.as_ref()),
-                expires: k
-                    .expire_at
-                    .map(|t| t.to_rfc3339())
-                    .unwrap_or_else(|| "never".into()),
-            })
-            .collect();
-        println!("{}", Table::new(rows).with(Style::markdown()));
-    }
-    Ok(())
-}
-
-/// What a `key create` response says about the key's credentials.
-#[derive(Debug)]
-enum KeyCreateMaterial<'a> {
-    /// The API returned the generated pair; it is shown once and never again.
-    Generated {
-        key_id: &'a str,
-        key_secret: &'a str,
-    },
-    /// The caller supplied pre-hashed credentials, so no pair is generated.
-    PreHashed,
-}
-
-/// Resolves the credentials to print from the request mode and the response.
-///
-/// Only a pre-hashed request explains a response without key material, so the
-/// mode is read from the request rather than inferred from what came back: an
-/// absent `keyId`/`keySecret` on a generated-key request means the one-time
-/// secret is lost, which is an error, not a "no key material returned" notice.
-fn resolve_key_create_material<'a>(
-    pre_hashed: bool,
-    key_id: Option<&'a str>,
-    key_secret: Option<&'a str>,
-    key_name: Option<&str>,
-) -> Result<KeyCreateMaterial<'a>, Box<dyn std::error::Error>> {
-    if pre_hashed {
-        return Ok(KeyCreateMaterial::PreHashed);
-    }
-    match (key_id, key_secret) {
-        (Some(key_id), Some(key_secret)) => Ok(KeyCreateMaterial::Generated { key_id, key_secret }),
-        _ => {
-            // Name the key when the response did return it, so the user knows
-            // which one to look for.
-            let named = match key_name {
-                Some(name) => format!(" '{}'", name),
-                None => String::new(),
-            };
-            Err(format!(
-                "the API response omitted the generated key material, so the one-time key secret \
-                 cannot be shown: the key{} may still have been created — list the organization's \
-                 keys and delete it if so",
-                named
-            )
-            .into())
-        }
-    }
-}
-
-pub async fn key_create(
-    client: &CloudClient,
-    opts: KeyCreateOptions,
-    json: bool,
-) -> Result<(), Box<dyn std::error::Error>> {
-    // Validate input before any network call so typos like --state broken
-    // fail locally instead of on the /organizations lookup.
-    let request = build_api_key_create_request(&opts)?;
-    let org_id = resolve_org_id(client, opts.org_id.as_deref()).await?;
-
-    let resp = client.create_api_key(&org_id, &request).await?;
-
-    let name = resp.key.as_ref().and_then(|key| key.name.as_deref());
-    // Resolve before either output branch, so --json cannot report success
-    // over a response that dropped the one-time key material.
-    let material = resolve_key_create_material(
-        request.hash_data.is_some(),
-        resp.key_id.as_deref(),
-        resp.key_secret.as_deref(),
-        name,
-    )?;
-
-    if json {
-        println!("{}", serde_json::to_string_pretty(&resp)?);
-    } else {
-        println!("API key created!");
-        println!("  Name: {}", or_absent(name));
-        match material {
-            KeyCreateMaterial::Generated { key_id, key_secret } => {
-                println!("  Key ID: {}", key_id);
-                println!("  Key Secret: {}", key_secret);
-                println!();
-                println!("Save the key secret now — it will not be shown again.");
-            }
-            KeyCreateMaterial::PreHashed => {
-                println!("  Pre-hashed credentials accepted; no generated key material returned");
-            }
-        }
-    }
-    Ok(())
-}
-
-pub async fn key_get(
-    client: &CloudClient,
-    key_id: &str,
-    org_id: Option<&str>,
-    json: bool,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let org_id = resolve_org_id(client, org_id).await?;
-
-    let key = client.get_api_key(&org_id, key_id).await?;
-
-    if json {
-        println!("{}", serde_json::to_string_pretty(&key)?);
-    } else {
-        print_human(&key)?;
-    }
-    Ok(())
-}
-
-pub async fn key_update(
-    client: &CloudClient,
-    key_id: &str,
-    opts: KeyUpdateOptions,
-    json: bool,
-) -> Result<(), Box<dyn std::error::Error>> {
-    // Validate input before any network call so typos like --state broken
-    // fail locally instead of on the /organizations lookup.
-    let request = build_api_key_update_request(&opts)?;
-    let org_id = resolve_org_id(client, opts.org_id.as_deref()).await?;
-
-    let key = client.update_api_key(&org_id, key_id, &request).await?;
-
-    if json {
-        println!("{}", serde_json::to_string_pretty(&key)?);
-    } else {
-        println!("API key {} updated", or_absent(key.name.as_deref()));
-        println!("  ID: {}", or_absent(key.id));
-        println!("  State: {}", or_absent(key.state.as_ref()));
-    }
-    Ok(())
-}
-
-pub async fn key_delete(
-    client: &CloudClient,
-    key_id: &str,
-    org_id: Option<&str>,
-    json: bool,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let org_id = resolve_org_id(client, org_id).await?;
-
-    let response = client.delete_api_key(&org_id, key_id).await?;
-    if json {
-        println!("{}", serde_json::to_string_pretty(&response)?);
-    } else {
-        println!("API key {} deleted", key_id);
-    }
-    Ok(())
-}
-
-// =============================================================================
-// Phase 6 — Activity command handlers
-// =============================================================================
-
-pub async fn activity_list(
-    client: &CloudClient,
-    org_id: Option<&str>,
-    from_date: Option<&str>,
-    to_date: Option<&str>,
-    json: bool,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let org_id = resolve_org_id(client, org_id).await?;
-
-    let activities = client.list_activities(&org_id, from_date, to_date).await?;
-
-    if json {
-        println!("{}", serde_json::to_string_pretty(&activities)?);
-    } else {
-        if activities.is_empty() {
-            println!("No activities found");
-            return Ok(());
-        }
-        #[derive(Tabled)]
-        struct Row {
-            #[tabled(rename = "ID")]
-            id: String,
-            #[tabled(rename = "Type")]
-            activity_type: String,
-            #[tabled(rename = "Created")]
-            created: String,
-        }
-        let rows: Vec<Row> = activities
-            .into_iter()
-            .map(|a| Row {
-                id: or_absent(a.id.as_deref()),
-                activity_type: or_absent(a.r#type.as_ref()),
-                created: or_absent(a.created_at.map(|at| at.to_rfc3339())),
-            })
-            .collect();
-        println!("{}", Table::new(rows).with(Style::markdown()));
-    }
-    Ok(())
-}
-
-pub async fn activity_get(
-    client: &CloudClient,
-    activity_id: &str,
-    org_id: Option<&str>,
-    json: bool,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let org_id = resolve_org_id(client, org_id).await?;
-
-    let activity = client.get_activity(&org_id, activity_id).await?;
-
-    if json {
-        println!("{}", serde_json::to_string_pretty(&activity)?);
-    } else {
-        print_human(&activity)?;
-    }
-    Ok(())
-}
-
-// =============================================================================
-// Phase 6 — Backup Config command handlers
-// =============================================================================
-
-pub async fn backup_config_get(
-    client: &CloudClient,
-    service_id: &str,
-    org_id: Option<&str>,
-    json: bool,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let org_id = resolve_org_id(client, org_id).await?;
-
-    let config = client.get_backup_config(&org_id, service_id).await?;
-
-    if json {
-        println!("{}", serde_json::to_string_pretty(&config)?);
-    } else {
-        print_human(&config)?;
-    }
-    Ok(())
-}
-
-pub async fn backup_config_update(
-    client: &CloudClient,
-    service_id: &str,
-    opts: BackupConfigUpdateOptions,
-    json: bool,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let org_id = resolve_org_id(client, opts.org_id.as_deref()).await?;
-    let request = build_backup_config_update_request(&opts);
-
-    let config = client
-        .update_backup_config(&org_id, service_id, &request)
-        .await?;
-
-    if json {
-        println!("{}", serde_json::to_string_pretty(&config)?);
-    } else {
-        println!("Backup configuration updated for service {}", service_id);
-        println!(
-            "  Backup period: {} hours",
-            or_absent(config.backup_period_in_hours)
-        );
-        println!(
-            "  Retention: {} hours",
-            or_absent(config.backup_retention_period_in_hours)
-        );
-        println!(
-            "  Start time: {}",
-            or_absent(config.backup_start_time.as_deref())
-        );
-    }
-    Ok(())
-}
-
-fn format_bytes(bytes: f64) -> String {
-    const KB: f64 = 1024.0;
-    const MB: f64 = KB * 1024.0;
-    const GB: f64 = MB * 1024.0;
-
-    if bytes >= GB {
-        format!("{:.2} GB", bytes / GB)
-    } else if bytes >= MB {
-        format!("{:.2} MB", bytes / MB)
-    } else if bytes >= KB {
-        format!("{:.2} KB", bytes / KB)
-    } else {
-        format!("{} B", bytes)
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -3681,58 +3020,6 @@ mod tests {
     }
 
     #[test]
-    fn resolve_key_create_material_returns_the_generated_pair() {
-        let material =
-            resolve_key_create_material(false, Some("key-id"), Some("key-secret"), Some("ci"))
-                .unwrap();
-        match material {
-            KeyCreateMaterial::Generated { key_id, key_secret } => {
-                assert_eq!(key_id, "key-id");
-                assert_eq!(key_secret, "key-secret");
-            }
-            KeyCreateMaterial::PreHashed => panic!("expected the generated pair"),
-        }
-    }
-
-    #[test]
-    fn resolve_key_create_material_reports_pre_hashed_regardless_of_response() {
-        assert!(matches!(
-            resolve_key_create_material(true, None, None, Some("ci")).unwrap(),
-            KeyCreateMaterial::PreHashed
-        ));
-        // The mode comes from the request, so echoed material does not change it.
-        assert!(matches!(
-            resolve_key_create_material(true, Some("key-id"), None, None).unwrap(),
-            KeyCreateMaterial::PreHashed
-        ));
-    }
-
-    #[test]
-    fn resolve_key_create_material_fails_when_generated_material_is_absent() {
-        let err = resolve_key_create_material(false, None, Some("key-secret"), Some("ci"))
-            .unwrap_err()
-            .to_string();
-        assert!(
-            err.contains("omitted the generated key material"),
-            "unexpected error: {err}"
-        );
-        assert!(err.contains("the key 'ci' may still have been created"));
-
-        // An empty string is material the API did send, so it is not absent.
-        let material = resolve_key_create_material(false, Some(""), Some(""), Some("ci")).unwrap();
-        assert!(matches!(material, KeyCreateMaterial::Generated { .. }));
-
-        // Without a name the message stays grammatical.
-        let err = resolve_key_create_material(false, Some("key-id"), None, None)
-            .unwrap_err()
-            .to_string();
-        assert!(
-            err.contains("the key may still have been created"),
-            "unexpected error: {err}"
-        );
-    }
-
-    #[test]
     fn generation_is_requested_unless_a_password_hash_is_sent() {
         let request = |new_password_hash: Option<&str>, new_double_sha1_hash: Option<&str>| {
             ServicePasswordPatchRequest {
@@ -3971,74 +3258,7 @@ mod tests {
     }
 
     #[test]
-    fn build_api_key_requests_support_hashes_and_ip_allowlists() {
-        let create_opts = KeyCreateOptions {
-            name: "ci-key".to_string(),
-            role_ids: vec!["a1a2a3a4-b1b2-c1c2-d1d2-e1e2e3e4e5e6".to_string()],
-            expires_at: Some("2025-12-31T23:59:59Z".to_string()),
-            state: Some("enabled".to_string()),
-            ip_allow: vec!["10.0.0.0/8".to_string()],
-            hash_key_id: Some("id-hash".to_string()),
-            hash_key_id_suffix: Some("abcd".to_string()),
-            hash_key_secret: Some("secret-hash".to_string()),
-            org_id: None,
-        };
-        let create_request = build_api_key_create_request(&create_opts).unwrap();
-        let create_json = serde_json::to_value(&create_request).unwrap();
-        assert_eq!(create_json["hashData"]["keyIdHash"], "id-hash");
-        assert_eq!(create_json["ipAccessList"][0]["source"], "10.0.0.0/8");
-        assert_eq!(
-            create_json["assignedRoleIds"][0],
-            "a1a2a3a4-b1b2-c1c2-d1d2-e1e2e3e4e5e6"
-        );
-
-        let update_opts = KeyUpdateOptions {
-            name: Some("renamed".to_string()),
-            role_ids: vec!["a1a2a3a4-b1b2-c1c2-d1d2-e1e2e3e4e5e6".to_string()],
-            expires_at: Some("2025-01-01T00:00:00Z".to_string()),
-            state: Some("disabled".to_string()),
-            ip_allow: vec!["0.0.0.0/0".to_string()],
-            org_id: None,
-        };
-        let update_request = build_api_key_update_request(&update_opts).unwrap();
-        let update_json = serde_json::to_value(&update_request).unwrap();
-        assert_eq!(update_json["expireAt"], "2025-01-01T00:00:00Z");
-        assert_eq!(update_json["state"], "disabled");
-        assert_eq!(update_json["ipAccessList"][0]["source"], "0.0.0.0/0");
-    }
-
-    #[test]
-    fn build_api_key_create_request_rejects_invalid_uuid() {
-        let opts = KeyCreateOptions {
-            name: "ci-key".to_string(),
-            role_ids: vec!["not-a-uuid".to_string()],
-            ..Default::default()
-        };
-        let err = build_api_key_create_request(&opts).unwrap_err();
-        assert!(
-            err.to_string().contains("not-a-uuid"),
-            "error should mention the bad value: {}",
-            err
-        );
-    }
-
-    #[test]
-    fn build_api_key_create_request_rejects_invalid_expire_at() {
-        let opts = KeyCreateOptions {
-            name: "ci-key".to_string(),
-            expires_at: Some("next-tuesday".to_string()),
-            ..Default::default()
-        };
-        let err = build_api_key_create_request(&opts).unwrap_err();
-        assert!(
-            err.to_string().contains("next-tuesday"),
-            "error should mention the bad value: {}",
-            err
-        );
-    }
-
-    #[test]
-    fn build_org_and_backup_config_requests_match_tested_shapes() {
+    fn build_org_update_request_matches_tested_shape() {
         let org_opts = OrgUpdateOptions {
             name: Some("Updated Org".to_string()),
             remove_private_endpoints: vec![
@@ -4054,18 +3274,6 @@ mod tests {
             "aws"
         );
         assert_eq!(org_json["enableCoreDumps"], false);
-
-        let backup_opts = BackupConfigUpdateOptions {
-            backup_period_hours: Some(12),
-            backup_retention_period_hours: Some(336),
-            backup_start_time: Some("03:00".to_string()),
-            org_id: None,
-        };
-        let backup_request = build_backup_config_update_request(&backup_opts);
-        let backup_json = serde_json::to_value(&backup_request).unwrap();
-        assert_eq!(backup_json["backupPeriodInHours"], 12.0);
-        assert_eq!(backup_json["backupRetentionPeriodInHours"], 336.0);
-        assert_eq!(backup_json["backupStartTime"], "03:00");
     }
 
     // Regression tests: invalid enum values must be rejected by build_* functions
@@ -4360,35 +3568,6 @@ mod tests {
         let err = build_update_service_request(&opts).unwrap_err();
         assert!(
             err.to_string().contains("turbo"),
-            "error should mention the bad value: {}",
-            err
-        );
-    }
-
-    #[test]
-    fn build_api_key_create_request_rejects_invalid_state() {
-        let opts = KeyCreateOptions {
-            name: "ci-key".to_string(),
-            state: Some("broken".to_string()),
-            ..Default::default()
-        };
-        let err = build_api_key_create_request(&opts).unwrap_err();
-        assert!(
-            err.to_string().contains("broken"),
-            "error should mention the bad value: {}",
-            err
-        );
-    }
-
-    #[test]
-    fn build_api_key_update_request_rejects_invalid_state() {
-        let opts = KeyUpdateOptions {
-            state: Some("broken".to_string()),
-            ..Default::default()
-        };
-        let err = build_api_key_update_request(&opts).unwrap_err();
-        assert!(
-            err.to_string().contains("broken"),
             "error should mention the bad value: {}",
             err
         );
